pyproject.toml
[tool.poetry]
name = "repl-nix-uxt-careers-website"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "repl_nix_uxt_careers_website"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
mysql-connector-python = "^8.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
